+++ cmd/neofs-lens/internal/flags.go
@@ -8,6 +8,7 @@ const (
 	flagAddress    = "address"
 	flagEnginePath = "path"
 	flagOutFile    = "out"
+	flagConfigFile = "config"
 )
 
 // AddAddressFlag adds the address flag to the passed cobra command.
@@ -32,4 +33,12 @@ func AddOutputFileFlag(cmd *cobra.Command, v *string) {
 	cmd.Flags().StringVar(v, flagOutFile, "",
 		"File to save object payload")
 	_ = cmd.MarkFlagFilename(flagOutFile)
+}
+
+// AddConfigFileFlag adds the config file flag to the passed cobra command.
+func AddConfigFileFlag(cmd *cobra.Command, v *string) {
+	cmd.Flags().StringVar(v, flagConfigFile, "",
+		"Path to file with storage node config")
+	_ = cmd.MarkFlagFilename(flagConfigFile)
+	_ = cmd.MarkFlagRequired(flagConfigFile)
 }